Add ContainerManager.java, the fundamental class for executor monitor
ContainerManager.java is the basic class, used for the Executor Thread. Specifically, it contains methods that allows the end user to start and stop containers. There are also methods, used to empower the robustness of the program, making sure that the container that the user wants to start/stop exist and they are not already active/stopped.