Add ContainerMonitorHttp.java that monitors docker containers with http requests
In this class we have methods that print information about containers such as statistics. We use them in ContainerVisualization.java in order to plot graphs, for example regarding to cpu usage. There are also methods that format the json files, which contain the information about the containers. We use them so as to print the information in a user-friendly way