Modify CSV.java in order to store container metrics
The method writeContainerInfoToCsv() is used to store: Container Name, Container Id, IP Address, Mac Address, CPU Usage for every container that the user wants to get info for, in a file with the name of it and the extension Data.csv. Currently, the file is stored in the project file OnePercent.