Modify ContainerVisualization.java so as to solve complex thread issues
We had to rewrite some parts of ContainerVisualization.java because it had issues with the threads we are running. We had to adapt it to the new changes regarding to the monitor and menu threads. An nested thread had to be modified too. We 've also done necessary imports and fixed issues with closing the window of the graph