Modify getContainerStats() in ContainerMonitorHttp.java
We 've added a String parameter to the signature of the method. This paramater indicates which of the classes has called the method so as we can print the appropriate message. When the method is called in ContainerVisualization.java, the output message indicates that a diagram is going to be plotted. When it is called by CSV.java, the message indicates that the data will be stored in a csv file