Change the Graph.java in order to include the storage of the data in the database
In the database we store the datetime and the cpu usage per second as long as the plot is running. I add the field flag in order to stop the storage in database when the user closes the window with the plot. I also add the field end to make the program wait for the user to close the window and then ask him if he wants to run again.